UI changes in Dropdown component
